remove 2d/3d_exchange.c
